+++ src/app/app.component.html
@@ -14,7 +14,6 @@
 ~ limitations under the License.
 -->
 
-
 <app-navigation *ngIf="ready">
     <router-outlet></router-outlet>
 </app-navigation>
