+++ src/app/day/day.component.spec.ts
@@ -18,6 +18,7 @@ import {async, ComponentFixture, TestBed} from '@angular/core/testing';
 
 import {DayComponent} from './day.component';
 import {HttpClientTestingModule} from '@angular/common/http/testing';
+import {OAuthModule} from 'angular-oauth2-oidc';
 
 describe('DayComponent', () => {
     let component: DayComponent;
@@ -27,6 +28,7 @@ describe('DayComponent', () => {
         TestBed.configureTestingModule({
             imports: [
                 HttpClientTestingModule,
+                OAuthModule.forRoot()
             ],
             declarations: [DayComponent]
         })
